server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-dotenv
httpx
pydantic
pydantic-settings
PyJWT
cryptography
